+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## future
+### New features
+- Add support for album artist.
+
 ## v0.1.0
 - Support for ID3v1 and flac.
 - Partial support for ID3v2, mp4, ASF and RIFF.